Try to fix previous commit
git-svn-id: https://svn.remlab.net/svn/miredo/trunk@637 6046fb1f-dae2-0310-8472-869a43925389

+++ libteredo/teredo.h
@@ -83,7 +83,8 @@ static inline int
 in6_matches_teredo_client (const union teredo_addr *ip6, uint32_t ip,
 				uint16_t port)
 {
-	return !((ip ^ ip6->teredo.client_ip) | (port ^ ip6->teredo.client_port));
+	return !(((uint32_t)(ip ^ ip6->teredo.client_ip))
+		| ((uint16_t)(port ^ ip6->teredo.client_port)));
 }
 
 /*
